value y is dynamic testing is working

+++ cz_app/lib/widget/app/models/graph.dart
@@ -1,18 +1,18 @@
 class Graph {
-  int Id;
-  int Month;
-  int AmmountOfNewReferrals;
-  int AmmountOfApprovedReferrals;
+  double Id;
+  double Month;
+  double AmmountOfNewReferrals;
+  double AmmountOfApprovedReferrals;
 
   Graph(this.Id, this.Month, this.AmmountOfNewReferrals,
       this.AmmountOfApprovedReferrals);
 
   factory Graph.fromJson(Map<String, dynamic> json) {
     return Graph(
-      json['id'] as int,
-      json['month'] as int,
-      json['ammountOfNewReferrals'] as int,
-      json['ammountOfApprovedReferrals'] as int,
+      json['id'] as double,
+      json['month'] as double,
+      json['ammountOfNewReferrals'] as double,
+      json['ammountOfApprovedReferrals'] as double,
 
     );
   }
